requirements.txt
Babel==2.7.0
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.3
django-allauth==0.39.1
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==2.0
django-phonenumber-field==3.0.1
django-searchable-select==1.5.0
Faker==1.0.7
idna==2.8
oauthlib==3.0.1
phonenumbers==8.10.14
Pillow==6.0.0
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
urllib3==1.25.3
yapf==0.27.0
